Add logging module to handle C library diagnostics and improve error reporting
